rename plot modules
