+++ src/arviz_plots/plots/compare_plot.py
