+++ src/arviz_plots/plots/dist_plot.py
