+++ src/arviz_plots/plots/ecdf_plot.py
