+++ src/arviz_plots/plots/ess_plot.py
