+++ src/arviz_plots/plots/pava_calibration_plot.py
