+++ src/arviz_plots/plots/ppc_pit_plot.py
