+++ src/arviz_plots/plots/ppc_rootogram_plot.py
