+++ src/arviz_plots/plots/prior_posterior_plot.py
@@ -7,7 +7,7 @@ from arviz_base import extract, rcParams
 from xarray import concat
 
 from arviz_plots.plot_collection import PlotCollection
-from arviz_plots.plots.distplot import plot_dist
+from arviz_plots.plots.dist_plot import plot_dist
 from arviz_plots.plots.utils import process_group_variables_coords, set_figure_layout
