+++ src/arviz_plots/plots/rank_plot.py
