+++ src/arviz_plots/plots/ridge_plot.py
