+++ src/arviz_plots/plots/trace_plot.py
